+++ common/src/main/java/dev/boxadactle/boxlib/events/BoxHook.java
@@ -1,0 +1,9 @@
+package dev.boxadactle.boxlib.events;
+
+public interface BoxHook<T> {
+
+    void register(T listener);
+
+    T invoker();
+
+}