+++ common/src/main/java/dev/boxadactle/boxlib/keybind/KeybindHelper.java
@@ -1,11 +1,8 @@
 package dev.boxadactle.boxlib.keybind;
 
 import com.mojang.blaze3d.platform.InputConstants;
-import dev.architectury.injectables.annotations.ExpectPlatform;
-import dev.boxadactle.boxlib.keybind.KeybindingImpl;
 import dev.boxadactle.boxlib.mixin.KeybindAccessor;
 import net.minecraft.client.KeyMapping;
-import org.apache.commons.lang3.NotImplementedException;
 
 /**
  * The KeybindHelper class provides utility methods for working with keybindings.