+++ common/src/main/java/dev/boxadactle/boxlib/test/TestMod.java
@@ -17,19 +17,15 @@ import dev.boxadactle.boxlib.translate.Language;
 import dev.boxadactle.boxlib.translate.Translator;
 import dev.boxadactle.boxlib.util.GuiUtils;
 import dev.boxadactle.boxlib.util.ModLogger;
-import dev.boxadactle.boxlib.util.RenderUtils;
 import dev.boxadactle.boxlib.util.WorldUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.phys.AABB;
 import oshi.util.tuples.Pair;
 
 public class TestMod {
 
     public static final String MOD_NAME = "Boxlib Test Mod";
-
-    public static final String MOD_ID = "boxlibtestmod";
 
     public static final ModLogger LOGGER = new ModLogger(MOD_NAME);