+++ fabric/src/main/java/dev/boxadactle/boxlib/fabric/BoxLibModmenu.java
@@ -1,13 +1,14 @@
-package dev.boxadactle.boxlib.test.fabric;
+package dev.boxadactle.boxlib.fabric;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
+import dev.boxadactle.boxlib.core.ModConstants;
 import dev.boxadactle.boxlib.test.config.ExampleConfigScreen;
 
-public class ModMenuIntegration implements ModMenuApi {
+public class BoxLibModmenu implements ModMenuApi {
 
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ExampleConfigScreen::new;
+        return ModConstants.IS_DEVELOPMENT ? ExampleConfigScreen::new : (screen) -> null;
     }
 }