+++ forge/src/main/java/dev/boxadactle/boxlib/forge/BoxLibForge.java
@@ -7,14 +7,17 @@ import dev.boxadactle.boxlib.core.BoxLib;
 import dev.boxadactle.boxlib.core.ModConstants;
 import dev.boxadactle.boxlib.keybind.KeybindingImpl;
 import dev.boxadactle.boxlib.scheduling.Scheduling;
+import dev.boxadactle.boxlib.test.TestMod;
+import dev.boxadactle.boxlib.test.config.ExampleConfigScreen;
 import dev.boxadactle.boxlib.util.MouseUtils;
-import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.client.event.RegisterClientCommandsEvent;
 import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
 import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.listener.SubscribeEvent;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLLoader;
 
@@ -25,6 +28,14 @@ public class BoxLibForge {
         BoxLib.init();
 
         ModConstants.IS_DEVELOPMENT = !FMLLoader.isProduction();
+
+        if (ModConstants.IS_DEVELOPMENT) {
+            TestMod.init();
+
+            ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () ->
+                    new ConfigScreenHandler.ConfigScreenFactory((minecraft, screen) -> new ExampleConfigScreen(screen))
+            );
+        }
     }
 
     @Mod.EventBusSubscriber(modid = ModConstants.MOD_ID, value = Dist.CLIENT)
@@ -51,14 +62,11 @@ public class BoxLibForge {
             Scheduling.tick();
         }
 
-    }
-
-    @Mod.EventBusSubscriber(modid = ModConstants.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
-    public static class ModForgeEvents {
         @SubscribeEvent
         public static void registerKeybinds(RegisterKeyMappingsEvent e) {
             KeybindingImpl.register(e::register);
         }
+
     }
 
 }