+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Changelog
-- update for newer minecraft versions
+- update for newer minecraft version 1.21.9 & 10
+- Color picker screen cannot close on escape now
+- BHexField has been deprecated